GUI: Preparatory work for dictionary entries aggregate list view
The idea is to allow the dictionaries entries list view to contain the words of the dictionaries selected in the dictionaries list view.

This means that we need to be able to add or remove all the words of a dictionary inside the dictionaries entries list view while keeping the words of other dictionaries.

The challenge is that dictionaries may contain a lot of words. Adding a lot of words to a List is reasonably fast, but removing them may be very slow - at least with List.removeAll(Collection).

This commit introduces an AggregateList which provides a single view of several lists. It also introduces an observable companion class: ObservableAggregateList.