gui: Improve grid edition
In order to avoid having too many buttons in the toolbar, this commit introduces the notion of mode to the toolbar, with two modes:

* In the default mode, the toolbar contains only the buttons required to work on a grid whose size has been defined
* In the 'resize' mode contains only the buttons to resize the grid (add/delete row/column buttons)

Switching between the two modes is possible via a new 'Resize grid' toggle button, which is visible on both modes.

This commit also changes the theme of the icons used for add/delete row/columns buttons.