gui: Fix dictionary view corruption
Upon quick dictionary select/unselect, dictionary view may get corrupt,
with words of the unselected dictionary remaining in the displayed word
list.

What happens:

1. The entry get selected -> A task to retrieve the words is launched
   and a reservation is made in the view model to store the result;
2. The entry is unselected -> The reservation in the view model is
   removed. It is supposed to clean the associated words which fails
   with an NPE (caught) because words are not retrieved yet.
3. Words retrieval task finally finish, presenter adds the words to the
   view model without consideration for the dictionary being unselected.

This commit fixes the issue by checking if the entry is really selected
when trying to add words to the dictionary view model. If not, words are
discarded. A check is also added to avoid throwing the NPE mentioned
above.