ci: Use custom build image
Use a custom build image "opensuse/croiseur-build-platform:15.1", based on
openSUSE Tumbleweed.

Alpine images are small but use the musl libc. Native solvers compiled on these
images do not work out-of-the-box on a Linux box with a more traditional libc.

Also simplify build: Reduce to one build job for Linux. Later we'll add jobs
for other platforms.