solver-ginsberg: Add element removal by negative pattern in trie
Implementation is not great though: Nodes are not really removed, only
terminal flag is unset. Also, it is quite slow, so it unfortunately
does not make the trie a good candidate for the current word caches
which must be frequently updated.

Pushing the implementation though as it can serve as a basis for a
smarter implementation later.