dictionary-tools: Add an experimental tool to score dictionary
This is a naive experiment to score dictionary based on their
crossability.

Here the only criterion taken is the cumulated number of crossings
between words taken two-by-two; The score is this number.

Examples:

- UKACD (250,594 words): 179,143,017,418
- General British English (224,634 words): 135,580,058,256

Despite parallelization, computation is very long due to complexity of
O(n^2). UKACD took ~2h on my hardware, which is quite old though.