gui: Implement basic word suggestion
Pragmatically using a simple FilteredList in gui-view-model because it
is more straightforward (and probably faster) than to call croiseur
search service and replace the list.

A reason that would require to change the solution is that if the filter
takes too long and freezes the UI. This would require to work
asynchronously; croiseur search call would make sense then since there
is already infrastructure to do asynchronous calls to croiseur in
gui-controller.