gui: Binds selected slot unsolvable status to suggestion list emptiness
This allows to immediately see if a partially filled selected word has
no match.

Since unsolvable status is represented by a specific background colour
of the boxes, the way to represent the focused box has been changed from
setting a specific background color to setting a specific border
width. The colour of the border has been set to white for focused shaded
boxes (which fixes the problem of the shaded boxes having no visual
indication of being focused).

Implementation is a bit tricky because unsolvable status can be set
from several places (CrosswordGridViewModel for selected slot,
CrosswordTextField when resetting the grid, GuiSolverPresenter when
presenting the result). Adjustments - which may not be ideal - have been
made to avoid write errors in these places due to the unsolvable status
value being bound.