gui: Include padding in grid max size definition
With rework of the dictionaries pane, the central zone of the solver's
BorderPane is a SplitPane with the grid on the left and the dictionaries
on the right.

In terms of spacing, this lead to switch the margin applied to the
center of the BorderPane to a padding applied to the CrosswordGridPane -
we only want the spacing to apply to the grid.

This broke CrosswordGridPane bindings to keep the cells as squares:
Padding is included in the width/height of the node (while margin was
not).

This commit fixes the issue by taking into account padding in grid
bindings definitions.