gui: Shorten property value getter/setter names
Drop some get/set prefixes: It is shorter to write.

These JavaBeans conventions are useful when properties need to be
accessed by tools, but it is not the case here.

The only usage here would be when using SceneBuilder/IntelliJ to
bind/set a property via a FXML file, but there is only a few properties
used this way and no custom ones.