cli: Reword cli tests so that they look more BDDesque
Introduced a custom fluent API, very useless, very fun.

Joke aside, it does come with an improvement: All assertions are added
to a single context which is verified in one go using JUnit's
Assert.assertAll() - a failing assertion not preventing the others to be
run.

This is useful because when an issue occurs, the first assertion failure
is often not enough to understand what happens - typically it's good
to have the error output when a test fails for an assertion on normal
output.

The fluent API does help in the sense it masks the big assertAll()s from
the tests.