dictionary: Use SoftReference for word list cached in Lazy
Dictionary words are lazily read and cached by dictionary providers,
using the Lazy class. This works great except that every loaded
dictionary is kept in memory indefinitely, even if they are not
actively used any more.

This commit tweaks the lazy implementation to wrap the cached value
within a SoftReference, allowing the value to be garbage-collected
upon memory pressure.