gui: Implement go back to welcome screen from editor view
All the scenes are preloaded and attached as properties of the
application's main stage at application startup.

Then, upon switch view events, controllers retrieve the desired scene
from the stage properties and set it as the stage's scene.

Probably not the most efficient way to handle multiple scenes, but it
does work and is arguably simple.