tests: Fix locale setting
The setting of default locale to English was made via a @Before hook,
just as the deployment step. Both methods had the same priority, which
means that deployment step could run before the setting of default
locale - which effectively happened sometimes, resulting on failures
on systems whose locale is not English.

This commit adjusts the priority of the locale setting hook so that it
always executes before the deployment step.