solvers: Bundle native libraries for multiple OS into a single jar
This way the jar is really portable, provided that all native libraries
are available.

Alternate method - packaging a jar for each os and architecture - is
better from disk usage point of view but is more difficult to set up.
Tried to configure it by defining new Gradle configurations and using
variant selection mechanism, but got lost and fell back to this simpler
solution.