Apply a module-info sort order convention
For consistency, adopt the following order convention in module-info
files:

1. requires
2. uses
3. exports
4. opens
5. provides

For each section, packages are sorted alphabetically; java.* modules may
be listed at the end of the section like it is common usage for imports.

Order is inspired by the key order found in openSUSE RPM spec files
(BuildRequires, Requires, Provides). 3,4,5 are all kinds of Provides, so
decided without any specific reason that it will be that order.