clue-xd: Add clues to export
It is less trivial than import because each XdClue contains the clue
and the answer to the clue, i.e. it requires access to the grid
slot contents.

The code to build the slot contents from the grid have been copied and
adapted from gui-view-model's SlotsViewModel. It might be moved later to
common if needed elsewhere.