gui: Make SplitPane center part not resizable with parent
Since all parts of the SplitPane are now *not* resizable with parent,
they actually all resize with parent *but* without triggering the
strange resize issues that occur when none are resizable or when only
center part is resizable.

I would prefer if only the center pane could resize but given the resize
issues mentioned above, this is an acceptable compromise.